+++ apps/mobile/src/app/(drawer)/(tabs)/(home,search,profile,notifications,settings)/posts/[id]/index.tsx
@@ -181,6 +181,15 @@ export default function Screen() {
         <CommentCard
           collapsed={item.data.collapsed}
           comment={item.data}
+          onCollapse={() => {
+            if (!collapsibleComments) {
+              return
+            }
+
+            collapse({
+              commentId: item.data.id,
+            })
+          }}
           onPress={() => {
             if (!collapsibleComments) {
               return