+++ apps/mobile/src/app/(drawer)/(tabs)/(home,search,profile,notifications,settings)/settings/gestures.tsx
@@ -2,7 +2,10 @@ import { ScrollView } from 'react-native-gesture-handler'
 import { useTranslations } from 'use-intl'
 
 import { type GestureAction } from '~/components/common/gestures'
-import { GestureIcons } from '~/components/common/gestures/actions'
+import {
+  GestureColors,
+  GestureIcons,
+} from '~/components/common/gestures/actions'
 import { Icon } from '~/components/common/icon'
 import { Menu } from '~/components/common/menu'
 import { type MenuItemOption } from '~/components/common/menu/item/options'
@@ -40,14 +43,14 @@ export default function Screen() {
     'postRightShort',
   ])
 
-  const options: Array<MenuItemOption<GestureAction>> = [
+  const postOptions: Array<MenuItemOption<GestureAction>> = [
     {
       label: t('options.upvote'),
       right: (
         <Icon
           name={GestureIcons.upvote}
           uniProps={(theme) => ({
-            tintColor: theme.colors.orange.accent,
+            tintColor: theme.colors[GestureColors.upvote].accent,
           })}
         />
       ),
@@ -59,7 +62,7 @@ export default function Screen() {
         <Icon
           name={GestureIcons.downvote}
           uniProps={(theme) => ({
-            tintColor: theme.colors.violet.accent,
+            tintColor: theme.colors[GestureColors.downvote].accent,
           })}
         />
       ),
@@ -71,7 +74,7 @@ export default function Screen() {
         <Icon
           name={GestureIcons.reply}
           uniProps={(theme) => ({
-            tintColor: theme.colors.blue.accent,
+            tintColor: theme.colors[GestureColors.reply].accent,
           })}
         />
       ),
@@ -83,7 +86,7 @@ export default function Screen() {
         <Icon
           name={GestureIcons.save}
           uniProps={(theme) => ({
-            tintColor: theme.colors.green.accent,
+            tintColor: theme.colors[GestureColors.save].accent,
           })}
         />
       ),
@@ -95,7 +98,7 @@ export default function Screen() {
         <Icon
           name={GestureIcons.hide}
           uniProps={(theme) => ({
-            tintColor: theme.colors.red.accent,
+            tintColor: theme.colors[GestureColors.hide].accent,
           })}
         />
       ),
@@ -107,11 +110,27 @@ export default function Screen() {
         <Icon
           name={GestureIcons.share}
           uniProps={(theme) => ({
-            tintColor: theme.colors.accent.accent,
+            tintColor: theme.colors[GestureColors.share].accent,
           })}
         />
       ),
       value: 'share',
+    },
+  ]
+
+  const commentOptions: Array<MenuItemOption<GestureAction>> = [
+    ...postOptions,
+    {
+      label: t('options.collapse'),
+      right: (
+        <Icon
+          name={GestureIcons.collapse}
+          uniProps={(theme) => ({
+            tintColor: theme.colors[GestureColors.collapse].accent,
+          })}
+        />
+      ),
+      value: 'collapse',
     },
   ]
 
@@ -140,7 +159,7 @@ export default function Screen() {
                   postLeftShort: next as GestureAction,
                 })
               }}
-              options={options}
+              options={postOptions}
               value={postLeftShort}
             />
 
@@ -151,7 +170,7 @@ export default function Screen() {
                   postLeftLong: next as GestureAction,
                 })
               }}
-              options={options}
+              options={postOptions}
               value={postLeftLong}
             />
           </>
@@ -177,7 +196,7 @@ export default function Screen() {
                   postRightShort: next as GestureAction,
                 })
               }}
-              options={options}
+              options={postOptions}
               value={postRightShort}
             />
 
@@ -188,7 +207,7 @@ export default function Screen() {
                   postRightLong: next as GestureAction,
                 })
               }}
-              options={options}
+              options={postOptions}
               value={postRightLong}
             />
           </>
@@ -218,7 +237,7 @@ export default function Screen() {
                   commentLeftShort: next as GestureAction,
                 })
               }}
-              options={options}
+              options={commentOptions}
               value={commentLeftShort}
             />
 
@@ -229,7 +248,7 @@ export default function Screen() {
                   commentLeftLong: next as GestureAction,
                 })
               }}
-              options={options}
+              options={commentOptions}
               value={commentLeftLong}
             />
           </>
@@ -255,7 +274,7 @@ export default function Screen() {
                   commentRightShort: next as GestureAction,
                 })
               }}
-              options={options}
+              options={commentOptions}
               value={commentRightShort}
             />
 
@@ -266,7 +285,7 @@ export default function Screen() {
                   commentRightLong: next as GestureAction,
                 })
               }}
-              options={options}
+              options={commentOptions}
               value={commentRightLong}
             />
           </>